Added a header background color and some more text
As I'm testing, I added a solid background color to the header section, so the background color should be black and the font color should be white. I noticed that there is a conflict with the rules, the color=black is overwriting the color=white so I'm gonna find out how to solve this.