Solved the text colors issues between the h1 general rule and the h1 header specific rule
I found out that if I need all of the h1 to share characteristics even between header, body or footer, but I need the h1 from some section to be different from the others but still share some other general rules, I must specify that difference with the specifity we studied yesterday. This way the h1 from the header are slightly different (white font color) from the rest of the h1, the same for the h2.