Major update: svg icons, section cursor style and button-alike behavior
I added a svg icon, there's still missing icons for each category but the implementation is done.

I also modified the entire CSS code and added some comments, I deleted .headerText specificity and included most of the properties the h1, h2 and h3 were sharing to the header itself.

I divided the main content in sections, you can even hover the cursor to make the hand cursor appear! and it's not only aesthetical, they are buttons.